refactor(web): 예약 폼 섹션을 ReservationFormContent로 변경
